Fotos produto

+++ E-commerce/src/main/java/org/serratec/ecommerce/api/service/ProdutoService.java
@@ -2,12 +2,15 @@ package org.serratec.ecommerce.api.service;
 
 import java.io.IOException;
 import java.net.URI;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
+import org.serratec.ecommerce.api.domain.Foto;
 import org.serratec.ecommerce.api.domain.Produto;
+import org.serratec.ecommerce.api.domain.dto.ProdutoDTO;
 import org.serratec.ecommerce.api.domain.dto.ProdutoInserirDTO;
+import org.serratec.ecommerce.api.exception.ProdutoNotFoundException;
 import org.serratec.ecommerce.api.repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -23,64 +26,62 @@ public class ProdutoService {
 	@Autowired
 	private FotoService fotoService;
 
-	public List<Produto> findAll() {
-		return prodRepo.findAll();
+	public List<ProdutoDTO> findAll() {
+		return prodRepo.findAll().stream().map(p -> adicionarImagemUri(p)).collect(Collectors.toList());
 	}
 
-	public Produto inserir(Produto produto) {
-		return prodRepo.save(produto);
-	}
-
-	public Optional<Produto> findById(Long id) {
-		return prodRepo.findById(id);
-	}
-
-	public void deleteById(Long id) {
-		prodRepo.deleteById(id);
-	}
-
-	public List<ProdutoInserirDTO> listar() {
-		List<Produto> produtos = prodRepo.findAll();
-		List<ProdutoInserirDTO>  dtos = new ArrayList<>();
-		for (Produto produto : produtos) {
-			ProdutoInserirDTO dto = adicionarImagemUri(produto);
-			dtos.add(dto);
+	public ProdutoDTO findById(Long id) throws ProdutoNotFoundException {
+		Optional<Produto> produtoDB = prodRepo.findById(id);
+		if (produtoDB.isEmpty()) {
+			throw new ProdutoNotFoundException(404, "Produto não encontrado!");
 		}
-		return dtos;
+
+		return adicionarImagemUri(produtoDB.get());
 	}
 
-	public ProdutoInserirDTO adicionarImagemUri(Produto produto) {
-		ProdutoInserirDTO dto = new ProdutoInserirDTO(produto);
-		dto.setNome(produto.getNomeProduto());
-		dto.setDescricao(produto.getDescricaoProduto());
-		dto.setQtdEstoque(produto.getQuantidadeEstoqueProduto());
-		dto.setValorUnitario(produto.getValorUnitarioProduto());
-		dto.setCategoria(produto.getCategoria());
-
-		URI uri = ServletUriComponentsBuilder.fromCurrentContextPath().path("/produtos/{id}/foto")
-				.buildAndExpand(produto.getId()).toUri();
-
-		dto.setUrlProduto(uri.toString());
-		return dto;
+	public ProdutoDTO inserir(ProdutoInserirDTO produto) {
+		return new ProdutoDTO(prodRepo.save(new Produto(produto)));
 	}
 
-	public ProdutoInserirDTO buscar(Long id) {
-		Optional<Produto> produto = prodRepo.findById(id);
-		return adicionarImagemUri(produto.get());
-	}
-
-	public ProdutoInserirDTO inserir(Produto produto, MultipartFile file) throws IOException {
-		produto = prodRepo.save(produto);
+	public ProdutoDTO inserir(ProdutoInserirDTO produtoDTO, MultipartFile file) throws IOException {
+		Produto produto = prodRepo.save(new Produto(produtoDTO));
 		fotoService.inserir(produto, file);
 		return adicionarImagemUri(produto);
 	}
 
-	public Produto findById(Produto produto) {
-		return prodRepo.findById(produto);
+	public void deleteById(Long id) throws ProdutoNotFoundException{
+		Optional<Foto> fotoDB = fotoService.buscarPorIdProduto(id);
+		if (fotoDB.isEmpty()) {
+			throw new ProdutoNotFoundException(404, "Produto não encontrado!");
+		}
+		fotoService.deleteFoto(fotoDB.get().getId());
+		prodRepo.deleteById(id);
+	}
+	
+	public ProdutoDTO update(Long id, ProdutoInserirDTO produto) throws ProdutoNotFoundException{
+		Optional<Produto> optional = prodRepo.findById(id);
+		if (optional.isEmpty()) {
+			throw new ProdutoNotFoundException(404, "Produto não encontrado");
+		}
+		Produto produtoDB = new Produto(produto);
+		produtoDB.setId(id);
+		ProdutoDTO novoProduto = new ProdutoDTO(prodRepo.save(produtoDB));
+		return novoProduto;
 	}
 
 	public Produto save(Produto produto) {
 		return prodRepo.save(produto);
 
 	}
+
+	public ProdutoDTO adicionarImagemUri(Produto produto) {
+		URI uri = ServletUriComponentsBuilder.fromCurrentContextPath().path("/produtos/{id}/foto")
+				.buildAndExpand(produto.getId()).toUri();
+
+		ProdutoDTO dto = new ProdutoDTO(produto);
+
+		dto.setUrlProduto(uri.toString());
+		return dto;
+	}
+
 }
